+++ raphodo/preferences.py
@@ -37,7 +37,7 @@ from raphodo.storage import (
 )
 from raphodo.generatenameconfig import *
 import raphodo.constants as constants
-from raphodo.constants import PresetPrefType
+from raphodo.constants import PresetPrefType, FileType
 from raphodo.utilities import available_cpu_count, make_internationalized_list
 import raphodo.__about__
 from raphodo.rpdfile import ALL_KNOWN_EXTENSIONS
@@ -542,7 +542,7 @@ class Preferences:
         """
         :return True if any of the pref lists contain a stored sequence no
         """
-        for pref_list in self.get_pref_lists():
+        for pref_list in self.get_pref_lists(file_name_only=True):
             if self._pref_list_uses_component(pref_list, STORED_SEQ_NUMBER):
                 return True
         return False
@@ -551,7 +551,7 @@ class Preferences:
         """
         :return True if any of the pref lists contain a session sequence no
         """
-        for pref_list in self.get_pref_lists():
+        for pref_list in self.get_pref_lists(file_name_only=True):
             if self._pref_list_uses_component(pref_list, SESSION_SEQ_NUMBER):
                 return True
         return False
@@ -560,7 +560,7 @@ class Preferences:
         """
         :return True if any of the pref lists contain a sequence letter
         """
-        for pref_list in self.get_pref_lists():
+        for pref_list in self.get_pref_lists(file_name_only=True):
             if self._pref_list_uses_component(pref_list, SEQUENCE_LETTER):
                 return True
         return False
@@ -716,12 +716,15 @@ class Preferences:
             v += s + "\n"
         return v
 
-    def get_pref_lists(self) -> Tuple[List[str], List[str], List[str], List[str]]:
+    def get_pref_lists(self, file_name_only: bool) -> Tuple[List[str], ...]:
         """
         :return: a tuple of the photo & video rename and subfolder
          generation preferences
         """
-        return self.photo_rename, self.photo_subfolder, self.video_rename, self.video_subfolder
+        if file_name_only:
+            return self.photo_rename, self.video_rename
+        else:
+            return self.photo_rename, self.photo_subfolder, self.video_rename, self.video_subfolder
 
     def get_day_start_qtime(self) -> QTime:
         """
@@ -754,8 +757,8 @@ class Preferences:
         else:
             return Qt.Unchecked
 
-    def pref_uses_job_code(self, pref_list: List[str]):
-        """ Returns True if the particular preferences contains a job code"""
+    def pref_uses_job_code(self, pref_list: List[str]) -> bool:
+        """ Returns True if the particular preference contains a job code"""
         for i in range(0, len(pref_list), 3):
             if pref_list[i] == JOB_CODE:
                 return True
@@ -763,7 +766,7 @@ class Preferences:
 
     def any_pref_uses_job_code(self) -> bool:
         """ Returns True if any of the preferences contain a job code"""
-        for pref_list in self.get_pref_lists():
+        for pref_list in self.get_pref_lists(file_name_only=False):
             if self.pref_uses_job_code(pref_list):
                 return True
         return False